+++ decidim-verifications/app/jobs/decidim/verifications/csv_census/application_job.rb
@@ -1,0 +1,10 @@
+# frozen_string_literal: true
+
+module Decidim
+  module Verifications
+    module CsvCensus
+      class ApplicationJob < ActiveJob::Base
+      end
+    end
+  end
+end