+++ decidim-verifications/spec/factories.rb
@@ -3,3 +3,4 @@
 require "decidim/core/test/factories"
 require "decidim/participatory_processes/test/factories"
 require "decidim/proposals/test/factories"
+require "decidim/verifications/test/factories"